Support some new features from the updated spec
Handle accessing section data at the top of the context via {{.}} and
add special-case handling for `null` sections. With these changes, all
test cases from the spec pass, except for optional new features that
are not yet implemented.

The specs submodule was updated in this PR only to verify this in CI as
well. It can be restored to point back to the fork prior to merge.